requirements.txt
flask
pydantic
fastapi
bson
pymongo[srv]
Flask-PyMongo
python-dotenv
flask-cors
flask-assets